+++ src/server/services/movie-api.ts
@@ -5,8 +5,6 @@ dotenv.config();
 const TMDB_KEY = process.env.TMDB_KEY || '';
 const TMDB_API_URL = 'https://api.themoviedb.org/3';
 
-//module.exports = () => {
-// eslint-disable-next-line import/prefer-default-export
 export const getGenres = async (mediatype: string) => {
   try {
     const requrl = `${TMDB_API_URL}/genre/${mediatype}/list`;
@@ -24,4 +22,23 @@ export const getGenres = async (mediatype: string) => {
     return err;
   }
 };
-//};
+
+export const discoverMedia = async (mediatype: string, genres: string, page: number) => {
+  try {
+    const requrl = `${TMDB_API_URL}/discover/${mediatype}`;
+    const media = await axios.get(requrl, {
+      params: {
+        api_key: TMDB_KEY,
+        language: 'en-US',
+        sort_by: 'popularity.desc',
+        with_genres: genres,
+        with_original_language: 'en',
+        page: page.toString(),
+      },
+    }).then((resp) => resp.data);
+    return media;
+  } catch (error) {
+    console.log(error);
+    return error;
+  }
+};
